Moved methods from slave_channels.go to slave.go
